+++ src/Card/ButtonHeader/ButtonHeader.js
@@ -1,26 +1,32 @@
 import React from 'react';
-import {bool, func, node, oneOf, string} from 'prop-types';
-import styles from './ButtonHeader.scss';
-import classNames from 'classnames';
+import {func, node, oneOf, string} from 'prop-types';
+
 import Button from '../../../src/Backoffice/Button';
 import WixComponent from '../../BaseComponents/WixComponent';
+import Header from '../Header';
+
+const buttonThemes = {
+  standard: 'whiteblueprimary',
+  emptyblue: 'emptyblue',
+  fullblue: 'fullblue'
+};
 
 class ButtonHeader extends WixComponent {
+  static displayName = 'ButtonHeader';
 
   static propTypes = {
+    ...Header.propTypes,
     title: node.isRequired,
     buttonTitle: string.isRequired,
     buttonOnClick: func.isRequired,
     buttonPrefix: node,
     buttonSuffix: node,
-    subtitle: node,
     tooltip: node,
     theme: oneOf([
       'standard',
       'fullblue',
       'emptyblue'
-    ]),
-    withoutDivider: bool
+    ])
   };
 
   static defaultProps = {
@@ -33,77 +39,43 @@ class ButtonHeader extends WixComponent {
   };
 
   render() {
-    const {title, subtitle, buttonOnClick, buttonTitle, buttonPrefix, buttonSuffix, withoutDivider, tooltip, theme} = this.props;
+    const {
+      title,
+      subtitle,
+      buttonOnClick,
+      buttonTitle,
+      buttonPrefix,
+      buttonSuffix,
+      withoutDivider,
+      tooltip,
+      theme
+    } = this.props;
 
-    const headerClasses = classNames({
-      [styles.headerOnlyTitle]: !subtitle,
-      [styles.headerTitleSubtitle]: subtitle,
-      [styles.withDivider]: !withoutDivider
-    });
-
-    const buttonClass = classNames({
-      [styles.button]: theme === 'standard',
-      [styles.buttonSmall]: theme !== 'standard'
-    });
-
-    const height = theme === 'standard' ? 'medium' : 'small';
-    let buttonTheme;
-    switch (theme) {
-      case 'standard': {
-        buttonTheme = 'whiteblueprimary';
-        break;
-      }
-      case 'emptyblue': {
-        buttonTheme = 'emptyblue';
-        break;
-      }
-      case 'fullblue':
-      default: {
-        buttonTheme = 'fullblue';
-        break;
-      }
-    }
     const buttonElement = (
-      <div className={buttonClass}>
+      <div>
         <Button
           dataHook="button"
-          height={height}
+          height={theme === 'standard' ? 'medium' : 'small'}
           suffixIcon={buttonSuffix}
           prefixIcon={buttonPrefix}
           onClick={buttonOnClick}
-          theme={buttonTheme}
-          >
-          {buttonTitle}
-        </Button>
+          theme={buttonThemes[theme] || buttonThemes.standard}
+          children={buttonTitle}
+          />
       </div>
     );
 
-    const tooltipElement = tooltip ? (
-      React.cloneElement(tooltip, {}, buttonElement)
-    ) : null;
-
-    const actionElement = tooltipElement ? tooltipElement : buttonElement;
-
-    const titleElement = (
-      <div data-hook="title" className={styles.title}>
-        {title}
-      </div>
-    );
-
-    const subtitleElement = subtitle ? (
-      <div data-hook="subtitle" className={styles.subtitle}>
-        {this.props.subtitle}
-      </div>
-    ) : null;
+    const tooltipElement = tooltip ?
+      React.cloneElement(tooltip, {children: buttonElement}) :
+      null;
 
     return (
-      <div className={headerClasses}>
-        <div>
-          {titleElement}
-          {subtitleElement}
-        </div>
-        {actionElement}
-      </div>
+      <Header
+        title={title}
+        subtitle={subtitle}
+        suffix={tooltipElement || buttonElement}
+        withoutDivider={withoutDivider}
+        />
     );
   }
 }