+++ src/Card/Divider/index.js
@@ -1,0 +1,10 @@
+import React from 'react';
+
+import styles from './styles.scss';
+
+const Divider = () =>
+  <div className={styles.root}/>;
+
+Divider.displayName = 'Divider';
+
+export default Divider;