+++ testkit/enzyme.js
@@ -168,6 +168,9 @@ export const fieldLabelAttributesTestkitFactory = enzymeTestkitFactoryCreator(fi
 import popoverMenuDriverFactory from '../src/PopoverMenu/PopoverMenu.driver';
 export const popoverMenuTestkitFactory = enzymeTestkitFactoryCreator(popoverMenuDriverFactory);
 
+import calendarDriverFactory from '../src/Calendar/Calendar.driver';
+export const calendarTestkitFactory = enzymeTestkitFactoryCreator(calendarDriverFactory);
+
 import datePickerFactory from '../src/DatePicker/DatePicker.driver';
 export const datePickerTestkitFactory = enzymeTestkitFactoryCreator(datePickerFactory);
 
@@ -209,6 +212,9 @@ export const headingTestkitFactory = enzymeTestkitFactoryCreator(headingDriverFa
 
 import textDriverFactory from '../src/Text/Text.driver';
 export const textTestkitFactory = enzymeTestkitFactoryCreator(textDriverFactory);
+
+import draggableDriverFactory from '../src/DragAndDrop/Draggable/Draggable.driver';
+export const draggableTestkitFactory = enzymeTestkitFactoryCreator(draggableDriverFactory);
 
 // wix-ui-backoffice proxy